Replaced connectivity monitor with php call and sms
Have commented the connectivity monitor code completely.  Before what we
were doing was, when the panic alert button was pressed, we checked if
connectivity monitor is online that if the phone is connected to the
internet, then we used the php scripts mechanism to send the panic
alert. If the connectivity monitor was offline, we sent the message
using normal phone sms. Now this was creating an issue especially using
connectivity monitor when deployed to ios, it was giving errors related
to cordova and got stuck when we clicked on panic alert. So instead,
what I've done is, comment out the connectivity monitor code altogether
and instead use direct php calling when sending panic alert. If mobile
is connected tot the internet, it will pass, otherwise it will fail. In
case of fail scenario, I have added the code to send sms using normal
sms procedure, if that also fails then we show an error to the user.
Also have added the building in cordova file for android. In this file
have set my own android_sdk path and jre path, but this would vary with
each dev environment.