Changes regarding the time zone
This commit includes changes for updating the timezone to pkr. In php
scripts the change has been done from CURRENT_TIMESTAMP to converting
the timestamp -5 hours that becomes pkr time. utc - 5, also in the other
web applicaiton project, the changes have been done for converting the
time when adding or retrieving to show pakistan time. Some changes still
need to be done at the database end, because in some tables, the column
type is timestamp, and in those tables, whenever a change occurs in any
column or row, the timestamp that is of that change gets set in that
row, so this causes a problem that timestamp isn't the correct one,
rather the last one. So, in case of registration, the registered date
time becomes equal to last updated time stamp. Should that column type
be even equal to timestamp. Will have to look into it, and ineed to
change the necessary data types from timestamp to datetime, which will
automatically fix this issue, and then even can revert the code from the
web application that is being checked in now.